Setup updates (#2783)
* No need to use shjs here (no?)

* This file is never used as the clean script copies it from the main app

* Centralized getting required npm/node versions and update npm version check script

* Extract some of the setup functions to a separate file

* Use addXmark like everywhere else

* Make repo deletion message clearer

* Add compare-versions to presetup script (#2776)

* Formatting

* Remove unnecessary npm version check script
