+++ internals/scripts/analyze.js
@@ -17,11 +17,10 @@ shelljs.exec(
 function callback() {
   clearInterval(progress);
   process.stdout.write(
-    '\n\nOpen ' +
-      chalk.magenta('http://webpack.github.io/analyse/') +
-      ' in your browser and upload the stats.json file!' +
-      chalk.blue(
-        '\n(Tip: ' + chalk.italic('CMD + double-click') + ' the link!)\n\n',
-      ),
+    `\n\nOpen ${chalk.magenta(
+      'http://webpack.github.io/analyse/',
+    )} in your browser and upload the stats.json file!${chalk.blue(
+      `\n(Tip: ${chalk.italic('CMD + double-click')} the link!)\n\n`,
+    )}`,
   );
 }
