--- internals/scripts/helpers/get-npm-config.js
@@ -1,3 +1,0 @@
-const fs = require('fs');
-
-module.exports = JSON.parse(fs.readFileSync('package.json', 'utf8'));
