+++ internals/scripts/helpers/get-required-node-npm-versions.js
@@ -1,0 +1,8 @@
+const {
+  engines: { node, npm },
+} = require('../../../package.json');
+
+module.exports = {
+  requiredNodeVersion: node.match(/([0-9.]+)/g)[0],
+  requiredNpmVersion: npm.match(/([0-9.]+)/g)[0],
+};
