+++ internals/scripts/setup.js
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
 
-const shell = require('shelljs');
 const { exec } = require('child_process');
 const path = require('path');
 const fs = require('fs');
@@ -11,7 +10,18 @@ const chalk = require('chalk');
 const animateProgress = require('./helpers/progress');
 const addCheckMark = require('./helpers/checkmark');
 const addXMark = require('./helpers/xmark');
-const npmConfig = require('./helpers/get-npm-config');
+const {
+  requiredNpmVersion,
+  requiredNodeVersion,
+} = require('./helpers/get-required-node-npm-versions');
+const {
+  initGitRepository,
+  addToGitRepository,
+  commitToGitRepository,
+  hasGitRepository,
+  checkIfRepositoryIsAClone,
+  removeGitRepository,
+} = require('./helpers/git-utils');
 
 process.stdin.resume();
 process.stdin.setEncoding('utf8');
@@ -32,70 +42,13 @@ function deleteFileInCurrentDir(file) {
 }
 
 /**
- * Checks if we are under Git version control
- * @returns {Promise<boolean>}
- */
-function hasGitRepository() {
-  return new Promise((resolve, reject) => {
-    exec('git status', (err, stdout) => {
-      if (err) {
-        reject(new Error(err));
-      }
-
-      const regex = new RegExp(/fatal:\s+Not\s+a\s+git\s+repository/, 'i');
-
-      /* eslint-disable-next-line no-unused-expressions */
-      regex.test(stdout) ? resolve(false) : resolve(true);
-    });
-  });
-}
-
-/**
- * Checks if this is a clone from our repo
- * @returns {Promise<any>}
- */
-function checkIfRepositoryIsAClone() {
-  return new Promise((resolve, reject) => {
-    exec('git remote -v', (err, stdout) => {
-      if (err) {
-        reject(new Error(err));
-      }
-
-      const isClonedRepo = stdout
-        .split(/\r?\n/)
-        .map(line => line.trim())
-        .filter(line => line.startsWith('origin'))
-        .filter(line => /react-boilerplate\/react-boilerplate\.git/.test(line))
-        .length;
-
-      resolve(!!isClonedRepo);
-    });
-  });
-}
-
-/**
- * Remove the current Git repository
- * @returns {Promise<any>}
- */
-function removeGitRepository() {
-  return new Promise((resolve, reject) => {
-    try {
-      shell.rm('-rf', '.git/');
-      resolve();
-    } catch (err) {
-      reject(err);
-    }
-  });
-}
-
-/**
  * Ask user if he wants to start with a new repository
  * @returns {Promise<boolean>}
  */
 function askUserIfWeShouldRemoveRepo() {
   return new Promise(resolve => {
     process.stdout.write(
-      '\nDo you want to start with a new repository? [Y/n] ',
+      '\nDo you want to delete the current git repository and create a new one? [Y/n] ',
     );
     process.stdin.resume();
     process.stdin.on('data', pData => {
@@ -161,7 +114,7 @@ function checkNodeVersion(minimalNodeVersion) {
       } else if (compareVersions(nodeVersion, minimalNodeVersion) === -1) {
         reject(
           new Error(
-            `You need Node.js v${minimalNodeVersion} or above but you have v${nodeVersion}`,
+            `You need Node v${minimalNodeVersion} or above but you have v${nodeVersion}`,
           ),
         );
       }
@@ -223,54 +176,6 @@ function installPackages() {
 }
 
 /**
- * Initialize a new Git repository
- * @returns {Promise<any>}
- */
-function initGitRepository() {
-  return new Promise((resolve, reject) => {
-    exec('git init', (err, stdout) => {
-      if (err) {
-        reject(new Error(err));
-      } else {
-        resolve(stdout);
-      }
-    });
-  });
-}
-
-/**
- * Add all files to the new repository
- * @returns {Promise<any>}
- */
-function addToGitRepository() {
-  return new Promise((resolve, reject) => {
-    exec('git add .', (err, stdout) => {
-      if (err) {
-        reject(new Error(err));
-      } else {
-        resolve(stdout);
-      }
-    });
-  });
-}
-
-/**
- * Initial Git commit
- * @returns {Promise<any>}
- */
-function commitToGitRepository() {
-  return new Promise((resolve, reject) => {
-    exec('git commit -m "Initial commit"', (err, stdout) => {
-      if (err) {
-        reject(new Error(err));
-      } else {
-        resolve(stdout);
-      }
-    });
-  });
-}
-
-/**
  * Report the the given error and exits the setup
  * @param {string} error
  */
@@ -299,17 +204,10 @@ function endProcess() {
 (async () => {
   const repoRemoved = await cleanCurrentRepository();
 
-  // Take the required Node and NPM version from package.json
-  const {
-    engines: { node, npm },
-  } = npmConfig;
-
-  const requiredNodeVersion = node.match(/([0-9.]+)/g)[0];
   await checkNodeVersion(requiredNodeVersion).catch(reason =>
     reportError(reason),
   );
 
-  const requiredNpmVersion = npm.match(/([0-9.]+)/g)[0];
   await checkNpmVersion(requiredNpmVersion).catch(reason =>
     reportError(reason),
   );
